+++ models/users.go
@@ -8,7 +8,7 @@ import (
 	"gopkg.in/yaml.v3"
 )
 
-// Lecture du fichier users.yaml et chargement dans la structure
+// Lecture du fichier beeusers.yaml et chargement dans la structure
 func LoadUsers() (err error) {
 
 	yamlFile, err := os.ReadFile(Config.UsersPath)
@@ -27,7 +27,7 @@ func LoadUsers() (err error) {
 	return err
 }
 
-// Lecture du fichier users.yaml et chargement dans la structure
+// Lecture du fichier beeusers.yaml et chargement dans la structure
 func GetUsersContent() (content string, err error) {
 
 	buf, err := os.ReadFile(Config.UsersPath)
@@ -39,7 +39,7 @@ func GetUsersContent() (content string, err error) {
 	return string(buf), err
 }
 
-// Mise à jour du fichiers users.yaml et rechargement dans la structure
+// Mise à jour du fichiers beeusers.yaml et rechargement dans la structure
 func UpdateUsers(content []byte) (err error) {
 
 	err = os.WriteFile(Config.UsersPath, content, 0644)
@@ -65,4 +65,102 @@ func CheckUser(user_id, password string) bool {
 		logs.Error("Tentative Connexion [%s]/[%s]", user_id, password)
 	}
 	return false
+}
+
+func (config *BeeConfig) IsUserAdmin(user_id string) bool {
+
+	if access, ok := config.Users[user_id]; ok {
+		return access.IsAdmin
+	}
+
+	return false
+}
+
+func (beeDir *BeeDir) IsUserEditor(user_id string) bool {
+
+	if Config.IsUserAdmin(user_id) {
+		return true
+	}
+
+	if beeDir.ParentID != "" {
+		bdir := Config.BeeDirs[beeDir.ParentID]
+		if access, ok := bdir.Users[user_id]; ok {
+			return access.IsEditor
+		}
+	} else {
+		if access, ok := beeDir.Users[user_id]; ok {
+			return access.IsEditor
+		}
+	}
+	return false
+}
+
+func (beeDir *BeeDir) IsUserReader(user_id string) bool {
+
+	if Config.IsUserAdmin(user_id) {
+		return true
+	}
+
+	if beeDir.ParentID != "" {
+		bdir := Config.BeeDirs[beeDir.ParentID]
+		if _, ok := bdir.Users[user_id]; ok {
+			return true
+		}
+	} else {
+		if _, ok := beeDir.Users[user_id]; ok {
+			return true
+		}
+	}
+
+	return false
+}
+
+// Mise à jour du fichiers .beeaccess.yaml et rechargement dans la structure
+func (beeDir *BeeDir) UpdateAccess(content []byte) (err error) {
+	pathAccess := beeDir.Path + "/.beeaccess.yaml"
+	err = os.WriteFile(pathAccess, content, 0644)
+	if err != nil {
+		msg := fmt.Sprintf("%s : [%v]", pathAccess, err)
+		logs.Error("WriteFile", msg)
+		return err
+	}
+	err = beeDir.LoadAccess()
+	if err != nil {
+		msg := fmt.Sprintf("%s : [%v]", Config.UsersPath, err)
+		logs.Error("LoadAccess", msg)
+		return err
+	}
+
+	return err
+}
+
+// Lecture du fichier beeusers.yaml et chargement dans la structure
+func (beeDir *BeeDir) GetAccessContent() (content string, err error) {
+	pathAccess := beeDir.Path + "/.beeaccess.yaml"
+	buf, err := os.ReadFile(pathAccess)
+	if err != nil {
+		// lecture du modèle
+		buf, _ := os.ReadFile("./conf/beeaccess-exemple.yaml")
+		return string(buf), nil
+	}
+	return string(buf), nil
+}
+
+// Lecture du fichier beeaccess.yaml et chargement dans la structure beedir s'il existe
+func (beeDir *BeeDir) LoadAccess() (err error) {
+	pathAccess := beeDir.Path + "/.beeaccess.yaml"
+
+	buf, err := os.ReadFile(pathAccess)
+	if err != nil {
+		// n'existe pas ou erreur
+		return nil
+	}
+
+	err = yaml.Unmarshal(buf, &beeDir.Users)
+	if err != nil {
+		msg := fmt.Sprintf("%s : [%v]", pathAccess, err)
+		logs.Error("Unmarshal", msg)
+		return err
+	}
+	return nil
 }