+++ routers/router.go
@@ -4,32 +4,44 @@ import (
 	"fmt"
 
 	"github.com/beego/beego/v2/server/web"
-	"github.com/beego/beego/v2/server/web/filter/auth"
 	"github.com/pbillerot/beerama/controllers"
 )
 
 func init() {
 
 	// Authentification obligatoire pour toutes les routes
-	filter := auth.NewBasicAuthenticator(controllers.SecretAuth, "Basic Authentificator")
-	web.InsertFilter("*", web.BeforeRouter, filter)
+	// filter := auth.NewBasicAuthenticator(controllers.SecretAuth, "Basic Authentification")
+	// web.InsertFilter("*", web.BeforeRouter, filter)
+	web.InsertFilter("*", web.BeforeExec, controllers.BasicAuthFilter)
 
 	// Routes sans rôle particulier
 	web.Router("/", &controllers.MainController{}, "get:Main")
-	web.Router("/folder/:beedirid", &controllers.MainController{}, "get:Folder")
 	web.Router("/return", &controllers.MainController{}, "get:Return")
-	web.Router("/folder/:beedirid/:htagid", &controllers.MainController{}, "get:FolderHtag")
-	web.Router("/search/:beedirid", &controllers.MainController{}, "post:Search")
 
 	// Routes static pour protéger les photos des albums
 	nsStatic := web.NewNamespace("/s",
 		web.NSBefore(controllers.AuthRequiredProfile),
 	)
 
+	// Routes avec rôle reader
+	nsReaderFolder := web.NewNamespace("/folder",
+		// Enchaîne plusieurs filtres : l'utilisateur doit être authentifié ET Editor
+		web.NSBefore(controllers.AuthRequiredProfile, controllers.ReaderRoleProfile),
+		web.NSRouter("/:beedirid", &controllers.MainController{}, "get:Folder"),
+		web.NSRouter("/download/:beedirid", &controllers.MainController{}, "post:FolderDownload"),
+		web.NSRouter("/:beedirid/:htagid", &controllers.MainController{}, "get:FolderHtag"),
+		web.NSRouter("/search/:beedirid", &controllers.MainController{}, "post:Search"),
+	)
+	nsReader := web.NewNamespace("/search",
+		// Enchaîne plusieurs filtres : l'utilisateur doit être authentifié ET Editor
+		web.NSBefore(controllers.AuthRequiredProfile, controllers.ReaderRoleProfile),
+		web.NSRouter("/:beedirid", &controllers.MainController{}, "post:Search"),
+	)
 	// Routes avec rôle editor
 	nsEditor := web.NewNamespace("/e",
 		// Enchaîne plusieurs filtres : l'utilisateur doit être authentifié ET Editor
 		web.NSBefore(controllers.AuthRequiredProfile, controllers.EditorRoleProfile),
+		web.NSRouter("/rename/:beedirid/:beefileid", &controllers.MainController{}, "post:FileRename"),
 		web.NSRouter("/rename/:beedirid", &controllers.MainController{}, "post:FolderRename"),
 		web.NSRouter("/reload/:beedirid", &controllers.MainController{}, "get:Reload"),
 		web.NSRouter("/meta/:beedirid/:beefileid", &controllers.MainController{}, "get:Meta;post:Meta"),
@@ -50,13 +62,14 @@ func init() {
 	nsAdmin := web.NewNamespace("/a",
 		// Enchaîne plusieurs filtres : l'utilisateur doit être authentifié ET Admin
 		web.NSBefore(controllers.AuthRequiredProfile, controllers.AdminRoleProfile),
+		web.NSRouter("/access/:beedirid", &controllers.MainController{}, "get:Access;post:Access"),
 		web.NSRouter("/mkdir", &controllers.MainController{}, "post:MkFolder"),
-		web.NSRouter("/users", &controllers.MainController{}, "get:Users;post:Users"),
 		web.NSRouter("/reload", &controllers.MainController{}, "get:ReloadAll"),
+		web.NSRouter("/users", &controllers.MainController{}, "get:Users;post:Users"),
 	)
 
 	// Ajouter les Namespaces au routeur Beego
-	web.AddNamespace(nsStatic, nsEditor, nsAdmin)
+	web.AddNamespace(nsStatic, nsReader, nsReaderFolder, nsEditor, nsAdmin)
 
 	fmt.Println("Routes init. Proceeding.")
 }